Changed the project structure, and also changed saving data to mognoDb into overwriting instead of skipping in case of duplicate ids
