+++ Dockerfile
@@ -1,17 +1,17 @@
 FROM python:3.9-slim
 
-RUN apt update
-RUN apt install -y python3-pip
-RUN apt clean
-RUN rm -rf /var/lib/apt/lists/*
-
-COPY requirements.txt /app/
-RUN pip3 install -r /app/requirements.txt
+RUN apt update && apt install -y python3-pip && apt clean && rm -rf /var/lib/apt/lists/*
 
 WORKDIR /app
 
-EXPOSE 8008
+COPY requirements.txt /app/
+RUN pip3 install -r requirements.txt
 
-COPY . .
+COPY src /app/src
+WORKDIR /app/src
+
+ENV PYTHONPATH="/app/src"
+
+EXPOSE 8008
 
 CMD ["waitress-serve", "--host=0.0.0.0", "--port=8008", "app:app"]
