+++ src/formatters.py
@@ -1,22 +1,4 @@
-import logging
-from staticIdsByCode import teamIdByCode
-
-def deletePropsFromStruct(data, delete):
-    for item in delete:
-        if item in data:
-            del data[item]
-
-    return data
-
-def drillForProp(data, parent, child):
-    if parent in data:
-        childVal = data[parent].pop(child, None)
-        data.pop(parent, None)
-    
-    if childVal:
-        data[child] = childVal
-
-    return data
+from static.IdsByCode import teamIdByCode
 
 propsToDelete = [  
                 "league",
@@ -32,21 +14,31 @@ propsToDelete = [
                 ]
 
 def formatGames(response):
+
     games = []
-    
     for item in response:
+
         item["_id"] = item["id"]
         item["homeId"] = teamIdByCode[item["teams"]["home"]["code"]]
         item["visitorId"] = teamIdByCode[item["teams"]["visitors"]["code"]]
+        
+        item["dateStart"] = item["date"]["start"].split('T')[0]
+        item["timeStart"] = item["date"]["start"].split('T')[1]
+           
         item.pop("id", None)
         item.pop("teams", None)
-
-
+        item.pop("start", None)
+        item.pop("date", None)
+        
         formatted = deletePropsFromStruct(item, propsToDelete)
-        formatted = drillForProp(formatted, "date", "start")
   
-        # ZAMENI VISITORTEAM I HOMETEAM SA FUNCIJOM GET TEAMID
         games.append(formatted)
 
     return games
 
+def deletePropsFromStruct(data, delete):
+    for item in delete:
+        if item in data:
+            del data[item]
+
+    return data
