--- utils.py
@@ -1,6 +1,0 @@
-from datetime import datetime
-
-def getTodaysDate():
-    DTnow = datetime.now() 
-    DTformatted = DTnow.strftime("%Y-%m-%d")
-    return DTformatted
